Enable any upstream_host for proxy health check
Enable setting any host as upstream_host for health check. Currently the upstream_host is only applied if it is {host} - and then it is taken from Host value. When proxying it may be necessary to give a different upstream_host than what the proxy responds to - and it will also
be necessary to set the same hostname as the Host when performing health checks.